fix types and comments

+++ MPU9250/Drivers/dev/inc/mpu9250.h
@@ -31,7 +31,7 @@ typedef struct{
 typedef struct{
 	axis_t gyro;
 	axis_t accl;
-	uint16_t temp;
+	float temp;
 }MPU9250_t;
 
 /* Exposed function prototypes -----------------------------------------------*/
@@ -40,7 +40,7 @@ retType APP_MPU9250Read(uint8_t addr2read, uint8_t * pdata, uint16_t size);
 retType APP_MPU9250Init(void);
 retType APP_MPU9250ReadGyro(axis_t * gyro);
 retType APP_MPU9250ReadAccl(axis_t * accl);
-retType APP_MPU9250ReadTemp(uint16_t * temp);
+retType APP_MPU9250ReadTemp(float * temp);
 
 
 #endif /* DEV_INC_MPU9250_H_ */
